refactor: Move element utility functions into a single file
