--- src/content/utils/create-xul-element.ts
@@ -1,8 +1,0 @@
-const XUL_NS = 'http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul';
-
-export function createXULElement<Name extends keyof XUL.XULElementTagNameMap>(
-  doc: Document,
-  name: Name,
-) {
-  return doc.createElementNS(XUL_NS, name) as XUL.XULElementTagNameMap[Name];
-}
