--- src/content/utils/get-xul-element-by-id.ts
@@ -1,3 +1,0 @@
-export function getXULElementById<E extends XUL.XULElement>(id: string): E {
-  return document.getElementById(id) as unknown as E;
-}
