+++ src/content/utils/index.ts
@@ -1,12 +1,14 @@
 export { buildCollectionFullName } from './build-collection-full-name';
 export { chunkString } from './chunk-string';
-export { createHTMLElement } from './create-html-element';
-export { createXULElement } from './create-xul-element';
+export {
+  createHTMLElement,
+  createXULElement,
+  getXULElementById,
+} from './elements';
 export { getAllCollectionItems } from './get-all-collection-items';
 export { getDOMParser } from './get-dom-parser';
 export { getItemURL } from './get-item-url';
 export { getLocalizedString } from './get-localized-string';
-export { getXULElementById } from './get-xul-element-by-id';
 export { hasErrorStack } from './has-error-stack';
 export { isObject } from './is-object';
 export { keyValue } from './key-value';
